requirements.txt
discord.py
pytz
fastapi
uvicorn
